pyproject.toml
[project]
name = "lora-messenger"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "textual",
    "pyserial",
    "cryptography",
]

[project.optional-dependencies]
sdr = [
    "numpy",
    "scipy",
    "pyadi-iio",
]
dev = [
    "pytest",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
